+++ semcore/bulk-textarea/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.4.0] - 2025-08-29
+
+### Added
+
+- Controlled mode for errors.
+- `onImmediatelyChange` internal property.
+
+### Fixed
+
+- Incorrect paddings for line numbers starting from 1000 lines.
+
 ## [16.3.1] - 2025-08-29
 
 ### Changed